// extensions/music-assistant-controls/src/music-assistant/external-code/interfaces.ts
export enum AlbumType {
  ALBUM = "album",
  SINGLE = "single",
  COMPILATION = "compilation",
  EP = "ep",
  UNKNOWN = "unknown",
}
// ...
export interface ProviderMapping {
  // Model for a MediaItem's provider mapping details.
  item_id: string;
  provider_domain: string;
  provider_instance: string;
  available: boolean;
  // quality details (streamable content only)
  audio_format?: AudioFormat;
  // optional details to store provider specific details
  details?: string;
  // url = link to provider details page if exists
  url?: string;
}

export interface MediaItemLink {
  type: LinkType;
  url: string;
}

export interface MediaItemImage {
  type: ImageType;
  path: string;
  provider: string;
  remotely_accessible: boolean;
}

export interface MediaItemChapter {
  position: number;
  name: string;
  start: number;
  end?: number;
}

export interface MediaItemMetadata {
  description?: string;
  review?: string;
  explicit?: boolean;
  images?: MediaItemImage[];
  genres?: string[];
  mood?: string;
  style?: string;
  copyright?: string;
  lyrics?: string;
  lrc_lyrics?: string;
  label?: string;
  links?: MediaItemLink[];
  performers?: string[];
  preview?: string;
  replaygain?: number;
  popularity?: number;
  cache_checksum?: string;
  chapters?: MediaItemChapter[];
}

interface _MediaItemBase {
  item_id: string;
  provider: string;
  name: string;
  version?: string;
  sort_name?: string;
  uri: string;
  external_ids?: Array<[ExternalID, string]>;
  is_playable: boolean; // if the item is playable (can be used in play_media command)
  translation_key?: string; // an optional translation key identifier
  media_type: MediaType;
}

export interface MediaItem extends _MediaItemBase {
  provider_mappings: ProviderMapping[];
  metadata: MediaItemMetadata;
  favorite: boolean;
  position?: number; //required for playlist tracks, optional for all other
  timestamp_added: number;
  timestamp_modified: number;
}

export interface ItemMapping extends _MediaItemBase {
  available: boolean;
  image?: MediaItemImage;
}

export interface Artist extends MediaItem {}

export interface Album extends MediaItem {
  year?: number;
  artists: Array<ItemMapping | Artist>;
  album_type: AlbumType;
}

export interface Track extends MediaItem {
  duration: number;
  artists: Array<ItemMapping | Artist>;
  // album track only
  album: ItemMapping | Album;
  disc_number?: number;
  track_number?: number;
}